refactoring spatial smoothing fixtures
