+++ tests/defences/preprocessor/test_spatial_smoothing_pytorch.py
@@ -29,37 +29,37 @@ from tests.utils import ARTTestException
 logger = logging.getLogger(__name__)
 
 
-# TODO these fixtures are duplicates from test_spatial_smoothing needs refactoring
-@pytest.fixture
-def image_batch(channels_first):
-    """
-    Image fixture of shape NHWC and NCHW.
-    """
-    test_input = np.repeat(np.array(range(6)).reshape(6, 1), 24, axis=1).reshape((2, 3, 4, 6))
-    if not channels_first:
-        test_input = np.transpose(test_input, (0, 2, 3, 1))
-    test_output = test_input.copy()
-    return test_input, test_output
-
-
-@pytest.fixture
-def video_batch(channels_first):
-    """
-    Video fixture of shape NFHWC and NCFHW.
-    """
-    test_input = np.repeat(np.array(range(6)).reshape(6, 1), 24, axis=1).reshape((1, 3, 2, 4, 6))
-    if not channels_first:
-        test_input = np.transpose(test_input, (0, 2, 3, 4, 1))
-    test_output = test_input.copy()
-    return test_input, test_output
-
-
-@pytest.fixture
-def tabular_batch():
-    """
-    Create tabular data fixture of shape (batch_size, features).
-    """
-    return np.zeros((2, 4))
+# # TODO these fixtures are duplicates from test_spatial_smoothing needs refactoring
+# @pytest.fixture
+# def image_batch(channels_first):
+#     """
+#     Image fixture of shape NHWC and NCHW.
+#     """
+#     test_input = np.repeat(np.array(range(6)).reshape(6, 1), 24, axis=1).reshape((2, 3, 4, 6))
+#     if not channels_first:
+#         test_input = np.transpose(test_input, (0, 2, 3, 1))
+#     test_output = test_input.copy()
+#     return test_input, test_output
+#
+#
+# @pytest.fixture
+# def video_batch(channels_first):
+#     """
+#     Video fixture of shape NFHWC and NCFHW.
+#     """
+#     test_input = np.repeat(np.array(range(6)).reshape(6, 1), 24, axis=1).reshape((1, 3, 2, 4, 6))
+#     if not channels_first:
+#         test_input = np.transpose(test_input, (0, 2, 3, 4, 1))
+#     test_output = test_input.copy()
+#     return test_input, test_output
+#
+#
+# @pytest.fixture
+# def tabular_batch():
+#     """
+#     Create tabular data fixture of shape (batch_size, features).
+#     """
+#     return np.zeros((2, 4))
 
 
 @pytest.mark.xfail(
